Increaded usability, added validation

+++ users/views.py
@@ -7,21 +7,6 @@ from django.contrib.auth.models import User
 # Create your views here.
 
 
-# def register(request):
-#     # Register a new user
-#     if request.method != 'POST':
-#         form = RegForm()
-#
-#     else:
-#         form = RegForm(data=request.POST)
-#
-#         if form.is_valid():
-#             new_user = form.save()
-#             login(request, new_user)
-#             return redirect('book_keeping:index')
-#     context = {'form': form}
-#     return render(request, 'users/register.html', context)
-
 def register(request):
     """ Register a new user """
 
@@ -57,7 +42,8 @@ def register(request):
 def list_accounts(request):
     """ Lists user accounts so that they may be deleted """
     users = User.objects.order_by('username')
-    users_dict = {'users': users}
+    error = ''
+    users_dict = {'users': users, 'error': error}
     return render(request, 'users/user_list.html', users_dict)
 
 
@@ -67,15 +53,26 @@ def delete_account(request):
     if request.method != 'POST':  # This means they clicked the delete button
         form = DeleteAccountForm()
         username = request.GET['delete']
-        context = {'form': form, 'username': username}
+        context = {'form': form, 'username': username, }
         return render(request, 'users/delete_account.html', context)
     else:
-        form = DeleteAccountForm(request.POST) 
+        form = DeleteAccountForm(data=request.POST)
+        initial_user = User.objects.get(username=request.POST['delete'])
         if form.is_valid():
+            users = User.objects.order_by('username')
             data = request.POST.copy()
             username_delete = data.get('username')
-            User.objects.filter(username=username_delete).delete()
-            return redirect('book_keeping:admin_index')
+            print(username_delete)
+            print(initial_user.username)
+            if (username_delete == initial_user.username):
+                message = 'User was deleted.'
+                User.objects.filter(username=username_delete).delete()
+                users_dict = {'users': users, 'message': message}
+                return render(request, 'users/user_list.html', users_dict)
+            else:  # The usernames did not match
+                error = 'Usernames did not match. No user was deleted.'
+                users_dict = {'users': users, 'message': error}
+                return render(request, 'users/user_list.html', users_dict)
 
     context = {'form': form}
     return render(request, 'users/delete_account.html', context)
